Renamed Header.js and edited the names in the corresponding files
