+++ src/components/pages/Dashboard/RenderDashboard.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import { Header } from '../../common/Header';
+import { DashHeader } from '../../common/DashHeader';
 import CaseTable from '../../common/CaseTable';
 import RecentActivity from '../../common/RecentActivity';
 
 function RenderDashboard() {
   return (
     <div>
-      <Header />
+      <DashHeader />
       <CaseTable />
       <RecentActivity />
     </div>
